% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfoot.R
\name{makeCutCountBAM}
\alias{makeCutCountBAM}
\title{Make a cutcount data file from a BAM file}
\usage{
makeCutCountBAM(bamFile, refgenome = "mm9")
}
\arguments{
\item{refgenome}{Reference genome, eg) "mm9", "hg19"}

\item{bamfile}{Path to the input bam file}
}
\value{
Bedgraph filename
}
\description{
A shorter version of makeCutCountBAMWithName
}
