% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfoot.R
\name{makeCutCountBAMWithName}
\alias{makeCutCountBAMWithName}
\title{Make a cutcount data file from a BAM file}
\usage{
makeCutCountBAMWithName(bamFile, name = "", refgenome = "mm9",
  atac = F)
}
\arguments{
\item{name}{Name of the output bedgraph file}

\item{refgenome}{Reference genome, eg) "mm9", "hg19"}

\item{atac}{Logical  TRUE for ATAC-seq}

\item{bamfile}{Path to the input bam file}
}
\value{
Bedgraph filename
}
\description{
This function reads a bam file and convert it to a BedGraph file which
contains counts of cuts (DNase) or insertions on the genomic locations.
}
